LibWeb: Reconstruct session history descriptors outside the traversable
Move the helpers that turn the UI process's session history descriptors
into local entries and document states next to the descriptor
constructors in SessionHistoryEntry.cpp, and make the navigable-level
resolution (retained entries, child reconstruction ids, navigation API
entry lists) LocalNavigable members. The history executor and the
population continuation both need them, and neither owns a traversable.

No behavior change.
