+++ Libraries/LibWeb/HTML/LocalTraversableNavigable.cpp
@@ -191,153 +191,6 @@ bool LocalTraversableNavigable::is_top_level_traversable() const
     return parent() == nullptr;
 }
 
-struct SessionHistoryEntryReconstructionState {
-    HashMap<CrossProcessId, RefPtr<DocumentState>> document_states;
-};
-
-static Vector<NonnullRefPtr<SessionHistoryEntry>> retained_session_history_entries(LocalNavigable& navigable)
-{
-    Vector<NonnullRefPtr<SessionHistoryEntry>> entries;
-    auto append = [&](RefPtr<SessionHistoryEntry> entry) {
-        if (!entry)
-            return;
-        if (entries.find_if([&](auto const& candidate) {
-                return candidate.ptr() == entry.ptr();
-            })
-            != entries.end()) {
-            return;
-        }
-        entries.append(entry.release_nonnull());
-    };
-
-    append(navigable.current_session_history_entry());
-    append(navigable.active_session_history_entry());
-
-    if (auto window = navigable.active_window()) {
-        for (auto const& navigation_entry : window->navigation()->entries())
-            append(navigation_entry->session_history_entry());
-    }
-
-    return entries;
-}
-
-static void prepare_child_navigable_history_reconstruction(LocalNavigable& navigable, SessionHistoryDocumentStateDescriptor const& document_state_descriptor)
-{
-    Vector<Optional<CrossProcessId>> child_navigable_ids;
-    child_navigable_ids.ensure_capacity(document_state_descriptor.nested_histories.size());
-    for (auto const& nested_history : document_state_descriptor.nested_histories)
-        child_navigable_ids.unchecked_append(nested_history.id);
-
-    auto active_entry = navigable.active_session_history_entry();
-    auto active_document = navigable.active_document();
-    if (active_entry && active_document
-        && active_entry->document_state()->cross_process_id() == document_state_descriptor.id) {
-        auto child_navigables = active_document->document_tree_child_navigables();
-        if (child_navigables.size() == child_navigable_ids.size()) {
-            // FIXME: This is temporary glue for the current load-then-seed ordering.
-            //        A replacement WebContent process can create live child navigables
-            //        before the UI process sends its canonical session-history tree.
-            //        Now that nested history ids are canonical CrossProcessIds, the UI id
-            //        must win; retarget the already-created child to match it. The
-            //        longer-term model should avoid creating a distinct temporary id
-            //        for a child the UI process already knows about.
-            for (size_t i = 0; i < child_navigables.size(); ++i) {
-                auto canonical_id = *child_navigable_ids[i];
-                child_navigables[i]->set_id_for_session_history_reconstruction(canonical_id);
-                child_navigable_ids[i].clear();
-            }
-        }
-    }
-
-    navigable.set_child_navigable_history_reconstruction_ids(move(child_navigable_ids));
-}
-
-static void apply_session_history_entry_descriptor_from_ui_process(SessionHistoryEntry& entry, SessionHistoryEntryDescriptor& entry_descriptor)
-{
-    entry.set_url(move(entry_descriptor.url));
-    entry.set_step(static_cast<int>(entry_descriptor.step));
-    entry.set_classic_history_api_state(move(entry_descriptor.classic_history_api_state));
-    entry.set_navigation_api_state(move(entry_descriptor.navigation_api_state));
-    entry.set_navigation_api_key(move(entry_descriptor.navigation_api_key));
-    entry.set_navigation_api_id(move(entry_descriptor.navigation_api_id));
-    entry.set_scroll_restoration_mode(entry_descriptor.scroll_restoration_mode);
-    entry.set_scroll_position_data(move(entry_descriptor.scroll_position_data));
-}
-
-static void apply_session_history_document_state_descriptor_from_ui_process(DocumentState& document_state, SessionHistoryDocumentStateDescriptor const& document_state_descriptor)
-{
-    VERIFY(document_state.cross_process_id() == document_state_descriptor.id);
-    document_state.set_history_policy_container(document_state_descriptor.history_policy_container);
-    document_state.set_request_referrer(document_state_descriptor.request_referrer);
-    document_state.set_request_referrer_policy(document_state_descriptor.request_referrer_policy);
-    document_state.set_initiator_origin(document_state_descriptor.initiator_origin);
-    document_state.set_origin(document_state_descriptor.origin);
-    document_state.set_about_base_url(document_state_descriptor.about_base_url);
-    document_state.set_resource(document_state_descriptor.resource);
-    document_state.set_reload_pending(document_state_descriptor.reload_pending);
-    document_state.set_ever_populated(document_state_descriptor.ever_populated);
-    document_state.set_navigable_target_name(document_state_descriptor.navigable_target_name);
-}
-
-static RefPtr<DocumentState> get_or_create_document_state_from_ui_process(SessionHistoryDocumentStateDescriptor const& document_state_descriptor, SessionHistoryEntryReconstructionState& reconstruction_state)
-{
-    RefPtr<DocumentState> document_state;
-    if (auto existing_document_state = reconstruction_state.document_states.get(document_state_descriptor.id); existing_document_state.has_value())
-        document_state = *existing_document_state;
-
-    if (!document_state) {
-        document_state = DocumentState::create(document_state_descriptor.id);
-        reconstruction_state.document_states.set(document_state_descriptor.id, document_state);
-    }
-
-    apply_session_history_document_state_descriptor_from_ui_process(*document_state, document_state_descriptor);
-    return document_state;
-}
-
-static NonnullRefPtr<SessionHistoryEntry> create_session_history_entry_from_ui_process(SessionHistoryEntryDescriptor entry_descriptor, SessionHistoryEntryReconstructionState& reconstruction_state)
-{
-    auto entry = SessionHistoryEntry::create();
-    apply_session_history_entry_descriptor_from_ui_process(*entry, entry_descriptor);
-
-    auto document_state = get_or_create_document_state_from_ui_process(entry_descriptor.document_state, reconstruction_state);
-    VERIFY(document_state);
-    entry->set_document_state(move(document_state));
-    return entry;
-}
-
-enum class PrepareChildHistoryReconstruction {
-    No,
-    Yes,
-};
-
-static NonnullRefPtr<SessionHistoryEntry> resolve_local_session_history_entry(LocalNavigable& navigable, SessionHistoryEntryDescriptor entry_descriptor, PrepareChildHistoryReconstruction prepare_child_history_reconstruction)
-{
-    auto retained_entries = retained_session_history_entries(navigable);
-    auto target_identity = session_history_entry_identity(entry_descriptor);
-    for (auto& retained_entry : retained_entries) {
-        if (session_history_entry_identity(*retained_entry) == target_identity) {
-            apply_session_history_entry_descriptor_from_ui_process(*retained_entry, entry_descriptor);
-            apply_session_history_document_state_descriptor_from_ui_process(*retained_entry->document_state(), entry_descriptor.document_state);
-            if (prepare_child_history_reconstruction == PrepareChildHistoryReconstruction::Yes) {
-                prepare_child_navigable_history_reconstruction(navigable, entry_descriptor.document_state);
-            }
-            return retained_entry;
-        }
-    }
-
-    SessionHistoryEntryReconstructionState reconstruction_state;
-    for (auto const& retained_entry : retained_entries) {
-        auto document_state = retained_entry->document_state();
-        if (document_state)
-            reconstruction_state.document_states.set(document_state->cross_process_id(), document_state);
-    }
-
-    if (prepare_child_history_reconstruction == PrepareChildHistoryReconstruction::Yes) {
-        prepare_child_navigable_history_reconstruction(navigable, entry_descriptor.document_state);
-    }
-    return create_session_history_entry_from_ui_process(move(entry_descriptor), reconstruction_state);
-}
-
 static bool is_same_document_push_or_replace(Optional<Bindings::NavigationType> navigation_type, SessionHistoryEntry const& target_entry, Optional<UniqueNodeID> displayed_document_id)
 {
     if (navigation_type != Bindings::NavigationType::Push
@@ -448,7 +301,7 @@ bool LocalTraversableNavigable::route_child_created_during_history_reconstructio
     VERIFY(child.parent().ptr() == &parent);
 
     child.prepare_to_populate_reconstructed_history_entry(navigation.target_entry.navigation_api_key);
-    prepare_child_navigable_history_reconstruction(child, navigation.target_entry.document_state);
+    child.prepare_child_navigable_history_reconstruction(navigation.target_entry.document_state);
 
     auto source_snapshot_params = snapshot_source_snapshot_params(nullptr);
     auto request = NavigationPopulationRequest {
@@ -1344,9 +1197,7 @@ void LocalTraversableNavigable::run_ui_history_step_unload_cancelation_job(Cross
 {
     (void)operation_id;
 
-    auto target_entry = resolve_local_session_history_entry(
-        *this, move(target_entry_descriptor),
-        PrepareChildHistoryReconstruction::No);
+    auto target_entry = resolve_local_session_history_entry(move(target_entry_descriptor), PrepareChildHistoryReconstruction::No);
     if (user_involvement == UserNavigationInvolvement::BrowserUI
         && ongoing_navigation().has<Utf16String>()
         && target_entry == current_session_history_entry()
@@ -1451,11 +1302,9 @@ void LocalTraversableNavigable::run_ui_changing_navigable_history_job(CrossProce
 
         apply_session_history_entry_descriptor_from_ui_process(*local_target_entry, target_entry);
         apply_session_history_document_state_descriptor_from_ui_process(*document_state, target_entry.document_state);
-        prepare_child_navigable_history_reconstruction(*navigable, target_entry.document_state);
+        navigable->prepare_child_navigable_history_reconstruction(target_entry.document_state);
     } else {
-        local_target_entry = resolve_local_session_history_entry(
-            *navigable, move(target_entry),
-            PrepareChildHistoryReconstruction::Yes);
+        local_target_entry = navigable->resolve_local_session_history_entry(move(target_entry), PrepareChildHistoryReconstruction::Yes);
     }
     if (!local_target_entry) {
         on_complete->function()(ChangingNavigableHistoryStepJobDisposition::Stale, UnloadDisplayedDocument::No);
@@ -1500,45 +1349,6 @@ void LocalTraversableNavigable::run_ui_changing_navigable_history_job(CrossProce
         operation.claimed_navigables_awaiting_continuation.set(navigable_id);
 }
 
-static Vector<NonnullRefPtr<SessionHistoryEntry>> session_history_entries_for_navigation_api_from_ui_process(LocalNavigable& navigable, Vector<SessionHistoryEntryDescriptor> entry_descriptors)
-{
-    auto retained_entries = retained_session_history_entries(navigable);
-    SessionHistoryEntryReconstructionState reconstruction_state;
-    for (auto const& retained_entry : retained_entries) {
-        auto document_state = retained_entry->document_state();
-        if (document_state)
-            reconstruction_state.document_states.set(document_state->cross_process_id(), document_state);
-    }
-
-    Vector<NonnullRefPtr<SessionHistoryEntry>> entries;
-    entries.ensure_capacity(entry_descriptors.size());
-
-    for (auto& entry_descriptor : entry_descriptors) {
-        RefPtr<SessionHistoryEntry> local_entry;
-        auto entry_identity = session_history_entry_identity(entry_descriptor);
-        for (auto const& retained_entry : retained_entries) {
-            if (session_history_entry_identity(*retained_entry) == entry_identity) {
-                local_entry = retained_entry;
-                break;
-            }
-        }
-
-        if (local_entry) {
-            apply_session_history_entry_descriptor_from_ui_process(*local_entry, entry_descriptor);
-            apply_session_history_document_state_descriptor_from_ui_process(*local_entry->document_state(), entry_descriptor.document_state);
-            entries.append(local_entry.release_nonnull());
-            continue;
-        }
-
-        auto entry = SessionHistoryEntry::create();
-        apply_session_history_entry_descriptor_from_ui_process(*entry, entry_descriptor);
-        entry->set_document_state(get_or_create_document_state_from_ui_process(entry_descriptor.document_state, reconstruction_state));
-        entries.append(move(entry));
-    }
-
-    return entries;
-}
-
 void LocalTraversableNavigable::prepare_ui_changing_navigable_for_unload(CrossProcessId operation_id, CrossProcessId navigable_id, GC::Ref<GC::Function<void()>> on_complete)
 {
     auto* operation = page().history_executor().find_history_operation(operation_id);
@@ -1569,7 +1379,7 @@ void LocalTraversableNavigable::apply_ui_changing_navigable_continuation(CrossPr
 
     Vector<NonnullRefPtr<SessionHistoryEntry>> entries_for_navigation_api;
     if (auto navigable = local_navigable_with_id(navigable_id); navigable && !navigable->has_been_destroyed())
-        entries_for_navigation_api = session_history_entries_for_navigation_api_from_ui_process(*navigable, move(entry_descriptors_for_navigation_api));
+        entries_for_navigation_api = navigable->session_history_entries_for_navigation_api_from_ui_process(move(entry_descriptors_for_navigation_api));
 
     apply_changing_navigable_history_step_continuation_impl(
         *continuation,
